+++ changelog.md
@@ -1,5 +1,11 @@
 # JQuiz Changelog
 
+## 0.2
+
+- Added profile ability to delete pending quiz turns.
+- Added personal progress overview to profile.
+- Reworked leaderboard to be more encouraging.
+
 ## 0.1
 
 - Added prototypic user account management